requirements.txt
python-telegram-bot==21.4
google-generativeai>=0.7.2
aiohttp>=3.9.5
python-dotenv>=1.0.1
gTTS>=2.5.1
